# Remove unnecessary creation of UAConfig instances