--- a/uaclient/entitlements/__init__.py
+++ b/uaclient/entitlements/__init__.py
@@ -24,11 +24,12 @@ ENTITLEMENT_CLASSES = [
 ]  # type: List[Type[UAEntitlement]]
 
 
-def entitlement_factory(name: str):
+def entitlement_factory(cfg: UAConfig, name: str):
     """Returns a UAEntitlement class based on the provided name.
 
     The return type is Optional[Type[UAEntitlement]].
     It cannot be explicit because of the Python version on Xenial (3.5.2).
+    :param cfg: UAConfig instance
     :param name: The name of the entitlement to return
     :param not_found_okay: If True and no entitlement with the given name is
         found, then returns None.
@@ -36,21 +37,21 @@ def entitlement_factory(name: str):
         entitlement with the given name is found, then raises this error.
     """
     for entitlement in ENTITLEMENT_CLASSES:
-        if name in entitlement().valid_names:
+        if name in entitlement(cfg=cfg).valid_names:
             return entitlement
     raise EntitlementNotFoundError()
 
 
 def valid_services(
-    allow_beta: bool = False, all_names: bool = False
+    cfg: UAConfig, allow_beta: bool = False, all_names: bool = False
 ) -> List[str]:
     """Return a list of valid (non-beta) services.
 
-    @param allow_beta: if we should allow beta services to be marked as valid
-    @param all_names: if we should return all the names for a service instead
+    :param cfg: UAConfig instance
+    :param allow_beta: if we should allow beta services to be marked as valid
+    :param all_names: if we should return all the names for a service instead
         of just the presentation_name
     """
-    cfg = UAConfig()
     allow_beta_cfg = is_config_value_true(cfg.cfg, "features.allow_beta")
     allow_beta |= allow_beta_cfg
 
@@ -65,10 +66,13 @@ def valid_services(
     if all_names:
         names = []
         for entitlement in entitlements:
-            names.extend(entitlement().valid_names)
+            names.extend(entitlement(cfg=cfg).valid_names)
 
         return sorted(names)
 
     return sorted(
-        [entitlement().presentation_name for entitlement in entitlements]
+        [
+            entitlement(cfg=cfg).presentation_name
+            for entitlement in entitlements
+        ]
     )