@orndorffgrant 2022-04-04
Thank you for this!!
Before:
```
root@dev-x:~# truncate -s 0 /var/log/ubuntu-advantage.log 
root@dev-x:~# ua attach TOKEN
...
root@dev-x:~# wc -l /var/log/ubuntu-advantage.log
954 /var/log/ubuntu-advantage.log
```
After:
```
root@dev-x:~# truncate -s 0 /var/log/ubuntu-advantage.log 
root@dev-x:~# ua attach TOKEN
...
root@dev-x:~# wc -l /var/log/ubuntu-advantage.log
166 /var/log/ubuntu-advantage.log
```